Move useRepositionMenu into separate file
